requirements.txt
discord.py
requests
aiohttp
google-generativeai
python-dotenv
